refactor : ReservationArticle 수정

+++ member-api/src/main/java/com/kernelsquare/memberapi/domain/reservation_article/service/ReservationArticleService.java
@@ -1,5 +1,6 @@
 package com.kernelsquare.memberapi.domain.reservation_article.service;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
@@ -54,10 +55,11 @@ public class ReservationArticleService {
 	public CreateReservationArticleResponse createReservationArticle(
 		CreateReservationArticleRequest createReservationArticleRequest) {
 		Member member = memberRepository.findById(createReservationArticleRequest.memberId())
-			.orElseThrow(() -> new BusinessException(MemberErrorCode.MEMBER_NOT_FOUND));
+			.orElseThrow(() -> new BusinessException(ReservationArticleErrorCode.MEMBER_NOT_FOUND));
 
 		ReservationArticle reservationArticle = CreateReservationArticleRequest.toEntity(
 			createReservationArticleRequest, member);
+
 		ReservationArticle saveReservationArticle = reservationArticleRepository.save(reservationArticle);
 
 		// HashTag 저장
@@ -132,16 +134,31 @@ public class ReservationArticleService {
 
 		List<FindAllReservationArticleResponse> responsePages = pages.getContent().stream()
 			.map(article -> {
+				Boolean articleStatus = canIReservation(article.getStartTime());
 				Long fullCheck = reservationRepository.countByReservationArticleIdAndMemberIdIsNull(article.getId());
 				return FindAllReservationArticleResponse.of(
 					article.getMember(),
 					article,
+					articleStatus,
 					fullCheck
 				);
 			})
 			.toList();
 
 		return PageResponse.of(pagination, responsePages);
+	}
+
+	private Boolean canIReservation(LocalDateTime startTime) {
+		boolean check = false;
+		LocalDateTime currentTime = LocalDateTime.now();
+
+		LocalDate minStartDate = startTime.toLocalDate();
+		LocalDate currentDate = currentTime.toLocalDate();
+
+		if (currentDate.isAfter(minStartDate.minusDays(7)) && currentDate.isBefore(minStartDate.minusDays(1))) {
+			check = true;
+		}
+		return check;
 	}
 
 	@Transactional(readOnly = true)
@@ -197,10 +214,10 @@ public class ReservationArticleService {
 				}
 			}
 
-			// 수정된 해시태그 5개 넘는지 체크로직
+			// 수정된 해시태그 10개 넘는지 체크로직
 			long currentHashtagCount = hashtagRepository.countAllByReservationArticleId(postId);
 			long updatedHashtagCount = currentHashtagCount + addHashtagCount - removeHashtagCount;
-			if (updatedHashtagCount >= 5) {
+			if (updatedHashtagCount >= 10) {
 				throw new BusinessException(ReservationArticleErrorCode.TOO_MANY_HASHTAG);
 			}
 
@@ -303,8 +320,15 @@ public class ReservationArticleService {
 
 	@Transactional
 	public void deleteReservationArticle(Long postId) {
-		reservationArticleRepository.findById(postId)
+		ReservationArticle reservationArticle = reservationArticleRepository.findById(postId)
 			.orElseThrow(() -> new BusinessException(ReservationArticleErrorCode.RESERVATION_ARTICLE_NOT_FOUND));
+
+		LocalDate currentDate = LocalDateTime.now().toLocalDate();
+		LocalDate startDate = reservationArticle.getStartTime().toLocalDate();
+
+		if (!currentDate.isBefore(startDate.minusDays(7))) {
+			throw new BusinessException(ReservationArticleErrorCode.DELETE_ONLY_BEFORE_7DAYS);
+		}
 
 		reservationArticleRepository.deleteById(postId);
